[#377] Show debug level logs in debug mode
This level was lowered to INFO in the old times when the whole buffer
content was part of the DEBUG logs, making them un-readable. This is
not the case any longer, so the change can be reverted. DEBUG messages
include the actual request/response messages, which is vital to find
the root cause of a bug.