[#514] Register commands with a unique prefix
The LSP executeCommand process requires the commands to be first registered with
the client. Vscode has a single registry of commands, for the entire client, and
shuts down a server if it tries to register a duplicate command.

This diff ensures that all commands used are unique, by prepending the
underlying operating system process id to the command when registered and used.

Fixes #514