Use OTP's get_doc if available
EEP48 provides documentation for erlang application/modules.
OTP-23 implements EEP48 and provide the doc info if available.

If user uses an OTP version that provides the functionality and the
doc_chunks are available use that by default and if not fallback to
fetching the docs from src via 'erlang_ls' own implementation.

This is a foundation for further work where we should write an
doc_chunk to markdown renderer, when markdown is available in the
client.