Do not make giant erlang nodes mesh
By default erlang nodes will try to make a fully connected net of all
erlang nodes discovered. We don't want that 'erlang_ls' nodes interfer
with other testing nodes on the same machine.

Also avoids confusing mnesia when several different schemas might
exist in the same group of nodes. Mnesia wants the same schema on all
nodes in the group.

An alternative solution is to make the nodes hidden, but that
might make connecting and inspecting the nodes harder.

If we later want to create a mesh of nodes this needs to be
revised, maybe use global_groups?