Prioritize Uris
If a module have several Uris prioritize them.

Search for project related Uri first and prioritize modules in src dir.

When working OTP reposiory we get at least two uris per module, one
in src tree and one in otp release tree, when jumping to definition
we want to jump to the one in the src tree and not to the release tree.

Also some modules names are duplicated in tests dir, avoid these too.

Now that the indexing is parallel the order is random, so it's not
given where to go.