Make doc chunk search use els db instead of code path
Before this change any els hover request used only the path of
the running VM to find the doc chunks. This made using otp_path
to point out which OTP version to use not work for hover requests.

With this change the hover provider now looks using els db for the
.beam and .chunk file in the local filesystem.