Change mfa application to parse from module
OTP-24 changes the start of the anno for application to
be from the M in the MFA. So instead of relying on the
application anno, we dig out the module and take that
anno for both OTP-24 and earlier. That way we get the same
result no matter which version of the parser we use.