Fix els_text:range to return [Start...End)
The old version returned [Start...End] which worked out
because the only place where it is used is for selecting
`-spec` and for `-spec` we actually wanted to also include
the last . so it worked.

However, in dcc7c3199e4a the -spec was changed to include .
so this odd behaviour is no longer needed.