[#1070] Only set the cookie if the provided one is a valid one
This fixes a recently introduced regression. The `erlang:get_cookie/0`
was moved earlier, to a stage where distribution was not started
yet. Since that function returns `nocookie` in a system with no
distribution, that caused the atom `nocookie` to be erroneously used
as a cookie. A cleaner approach could be used, but that would require
a bit of refactoring. For now this should fix the regression with
minimal code changes.