[#1073] Add test to find non-parametrized macro definitions.
The code calls ?MACRO_B/2, but has only a ?MACRO_B definition without
parameters. As this actually works, erlang_ls should find the
definition, so let's test this.

The test changes the code_navigation.erl file, so a few other tests had
to be updated also to cope with this change.