Introduce diagnostic codes for the OTP compile
Expose the module and error passed to Module:print_error as an error code.

This allows us to provide telemetry of the codes, for later analysis, in
situations where this is practical.

It also allows an easy tie-up to any external documentation about how to deal
with the error.

The mechanism is to use the source of the original format_error/2 clauses to map
onto codes, which have a regular format, e.g. L1227 where the first letter
indicates the linter, and the rest is just a 4 digit unique number.

This format fits in with with existing error presentation in IDEs, as it is
already used by other compilers such as rust.