Send telemetry notification with compiler diagnostic codes
When enabled, default off, send a `telemetry/event` notification containing the
URI and error codes for compiler diagnostics in the given file.

The allows a client to log the telemetry, and perform later analysis on the
occurrence of diagnostics types.

Note: any logging takes place at the client level, if enabled, to whatever
location makes sense for the specific client.