Reload project-specific files on every Graudalizer diagnostic run
The check to determine if project files should be imported into Gradualizer
comes from the time when the integration relied on Gradualizer being available
through ERL_LIBS and would only be started by the diagnostic being called.

Now that Gradualizer is a dependency of ErlangLS,
it's running on each startup since ErlangLS itself.
This means the check never succeeds, therefore the project specific files
necessary for properly setting up the typechecker are never loaded.
This leads to false positives being reported such as missing remote types or
record definitions.

Reimporting the project files on each run might be a bit redundant,
but it seems to work reasonably well in practice.
Dogfooding this solutions shows it doesn't lead to visible unnecessary
load on the system.