Revert "Precompute list of enabled diagnostics" (#1197)
This reverts commit 2f6decd8cfe342da60e102b55cde66c4af7f784a.

This change introduced a regression in the debugger, preventing it from starting.
The `els_dap` escript invokes the `els_config` initialization procedure
(see https://github.com/erlang-ls/erlang_ls/commit/4b475b0864f433ccc08c80609478f5b15b9b41b7),
but the new version of the initialization depends on the `els_diagnostics` module,
which is part of the `els_lsp` application, not included in the `els_dap` escript.

For now simply reverting the change, but we should revisit the application structure.
The original idea was for an application to contain the implementation of the JSON-RPC protocol
and for the `els_dap` and `els_lsp` to utilize that as a dependency. That never really happened
and the current application split in Erlang LS in its current form feels a bit arbitrary.
One may argue that Erlang LS could get rid of the umbrella structure and be a single application.