Single Provider Process
Erlang LS historically used one process per provider. This was
problematic, since requests handled by different providers could lead
to race conditions, causing a server crash. For example, a completion
request could not see the latest version of the text in case of
incremental text synchronization.

There was very little gain in parallelizing providers, so we decided to
simplify the architecture and have a single process, instead. There's
still room for refactoring and cleanup, but those will be handled as a
follow up.

As part of the architectural change, we also dropped support for BSP
since it has been stalling for too long time. Instead, we will try to
implement a simpler solution which allows the extraction of paths from
the build system (e.g. via the rebar.conifg).

* Drop BSP Support
* Introduce text synchronization provider
* Run providers as a single process
* Remove buffer server, index in the background instead
* Decouple DAP provider from LSP one
* Brutally kill background jobs
* Set standard_io as default value for the io_device