Use version from didOpen, use text from editor as source of truth
Without this change, the server could read an outdated version of the text from disk and
try to apply changes to it. This was a very common scenario in case a file was already opened
in the editor with pending changes, not yet saved to disk. A reload of the server would cause
the server to incorrectly rely on the version on disk.

The version is also used now, so that eventual background jobs on outdated versions of the code
would abort.