Avoid race condition around POIs extraction
The deep_index function does not store the result of the indexing if a
new version of the document is present in the DB. This could lead to a
race condition where the caller of the ensure_deeply_index function
expects POIs to be expanded but, since the document is re-read from
the DB, they could still be un-expanded.

With this change, the ensure_deeply_index function returns a version
of the document with POIs always expanded, even if they may be
slightly outdated.