Track open files (#1292)
* Track open files

It can happen that some files are opened in the editor. If these files
have pending changes (unsaved), we should consider them as the "source
of truth" when it comes to text edits flowing from the IDE to the
language server.

In case of an external operation (e.g. a checkout or rebase in a
version control system), Erlang LS inconditionally reloads changed
files from disk. This is an incorrect behaviour, since subsequent
operations applied via the IDE to a non up-to-date version of the
buffer can result in errors. The language server should ignore such
external operations when the file is opened in the editor.

Eventual inconsistencies between the disk-copy and the IDE-copy of the
file will be resolved in the context of the IDE (e.g. in VS Code the
user is warned about the situation on save) and text edits from
the IDE can always be trusted.