Only store include/include_lib POIs as string
In case of incorrect syntax, it can happen that the content of an
include or include_lib is a tree and not a string. Since the
erl_syntax:string_value does not guarantee a string as the output,
let's explicitly validate the returned value. This will ensure that
non-string values are not stored in the DB, causing diagnostics to
fail due to wrong type assumptions.