Merge server and provider processes.
The els_provide process is a remain from the times where providers had
their own dedicated processes and it does not currently provide
additional value. It actually makes provider errors not bubble up to
the client. In fact, in case of a provider error the provider process
would crash, but the server would indefinitely wait for a
response. The error would not be returned via LSP.

This change simplifies the architecture of Erlang LS by merging the
els_server and els_provider processes:

* Get rid of els_provider process and internal_state
* Remove obsolete available_providers function
* Do not pass server state to providers

As a follow up, it should be possible to:

* Pass requests as-they-are (as opposed to only the parameters) to
providers, avoiding an un-necessary translation layer
* Generalize the code in els_method