Implement goto_definition for testcases
Summary:
https://github.com/erlang-ls/erlang_ls/issues/1205 suggested supporting the goto-definition feature for testcases.
Testcases are atoms (function definitions) and currently whenever doing goto-definition on an atom we treat it as a module.
I simply modified goto_definition to first look for a function of arity 1 in the current module when it is unclear whether the user is looking for a module or a testcase.

Test Plan:
I added the "testcase" test-case in els_definition_SUITE, following the structure of all other test-cases in that file.
This new testcase corresponds to doing "goto defintion" on the atom "one" which appears in "all() -> [ one ]" in sample_SUITE.erl.
It verifies that we are sent to the definition of one at the bottom of sample_SUITE.erl.
I then checked that this new test-case passed, and that I did not break any other test in that suite. In particular, there were already tests for goto-definition on module names, and they still pass.

Reviewers:
robertoaloi