Register capabilities dynamically, complete support for didChangedWatchedFiles
By being able to register capabilities dynamically, Erlang LS can
instruct the client to send `didChangeWatchedFiles`
requests whenever a file or directory is modified outside of the
client itself. Through this mechanism it is possible to prevent text
synchronization issues and crashes during events such as a rebase or a checkout.