Use explicit symbol range for document symbols
The previous implementation relied on the wrapping
range, which could contain zeros as column numbers. When transferring
the wrapping range via the LSP protocol, this could produce negative
numbers which are supported in some editors (eg Emacs) but not in
others (eg VS Code).

An alternative solution could have been refactor the wrapping range
handling, but for the time being - and considering symbol ranges are
limited to functions - I prefer to treat the two entities separately
for simplicity and to minimize risks.