[1371] Add Custom Hostname and Domain Options
When using `longnames` in projects that don't
adhere to the host provided hostname and domain
it is useful to be able to override those via the
`erlang_ls.config` file.  The code in this commit
provides these options as well as cleans up
some related duplicated code.

This code also fixes a small bug that would leave
a trailing dot `.` if a domain is not defined.

Tests have been updated to reflect these changes and
new tests have been created to test the new options.