Show function docs for function_clauses and specs
`els_docs:docs/2` was not previously implemented for `function_clause`
and `spec` POIs, so hover requests would return `null` when hovering
over function and spec definitions. This change returns the function
definition docs for both cases.

The new behavior is in line with other language servers (ElixirLS,
rust-analyzer) and can be useful - for example to see how the
documentation looks for a function you're just written.